requirements.txt
# Core dependencies
requests>=2.28.0
click>=8.0.0
pyyaml>=6.0
python-dotenv>=0.19.0

# Development tools
black>=22.0.0
flake8>=5.0.0
pytest>=7.0.0
pytest-cov>=4.0.0

# Utility libraries
rich>=12.0.0
typer>=0.7.0
pathlib2>=2.3.0

# Data processing
pandas>=1.5.0
numpy>=1.21.0

# Web framework (optional)
fastapi>=0.85.0
uvicorn>=0.18.0

# Testing
httpx>=0.23.0
pytest-asyncio>=0.20.0